Refactored wlm utils
